+++ application/src/main/java/com/musicforall/services/songlist/SonglistService.java
@@ -1,4 +1,4 @@
-package com.musicforall.services;
+package com.musicforall.services.songlist;
 
 import com.musicforall.model.Song;
 import com.musicforall.model.Songlist;
