+++ application/src/main/java/com/musicforall/web/IndexController.java
@@ -1,19 +1,11 @@
 package com.musicforall.web;
 
-import com.musicforall.model.Song;
-import com.musicforall.model.Songlist;
-import com.musicforall.model.Tag;
-import com.musicforall.model.User;
-import com.musicforall.services.SongService;
-import com.musicforall.services.SonglistService;
-import com.musicforall.services.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.web.csrf.CsrfToken;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
