+++ application/src/test/java/com/musicforall/SonglistServiceTest.java
@@ -3,9 +3,9 @@ package com.musicforall;
 import com.musicforall.model.Song;
 import com.musicforall.model.Songlist;
 import com.musicforall.model.User;
-import com.musicforall.services.SongService;
-import com.musicforall.services.SonglistService;
-import com.musicforall.services.UserService;
+import com.musicforall.services.song.SongService;
+import com.musicforall.services.songlist.SonglistService;
+import com.musicforall.services.user.UserService;
 import com.musicforall.util.JpaServicesTestConfig;
 import com.musicforall.util.ServicesTestConfig;
 import org.junit.After;
@@ -17,7 +17,6 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.support.AnnotationConfigContextLoader;
 
-import javax.validation.constraints.AssertTrue;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
